made kube warning component, ui changes
