+++ packages/manager/src/features/Kubernetes/kubeUtils.ts
@@ -6,8 +6,6 @@ import { LinodeType } from '@linode/api-v4/lib/linodes';
 import { pluralize } from 'src/utilities/pluralize';
 import { ExtendedCluster, ExtendedPoolNode, PoolNodeWithPrice } from './types';
 
-export const nodeWarning = `We recommend at least 3 nodes in each pool. Fewer nodes may affect availability. For more info see `;
-
 export const getMonthlyPrice = (
   type: string,
   count: number,
